requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.1.2
flask-sock==0.5.2
h11==0.13.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
simple-websocket==0.5.2
Werkzeug==2.1.2
wsproto==1.1.0
selenium~=4.3.0
requests~=2.28.1
pandas~=1.4.3
scipy~=1.8.1
numpy~=1.23.1
pyfiglet~=0.8.post1
python-dotenv~=0.20.0